refactor: keep foundation modules within source ratchet
