+++ src/usage-aggregator.ts
@@ -1,4 +1,3 @@
-import { homedir } from 'node:os'
 import { CATEGORY_LABELS, type ProjectSummary, type TaskCategory, type DateRange } from './types.js'
 import { type PeriodData, type ProviderCost, type MenubarPayload, type ClaudeConfigSelector, buildMenubarPayload } from './menubar-json.js'
 import { parseAllSessions, filterProjectsByName, filterProjectsByDays, filterProjectsByClaudeConfigSource, isSessionHydrationComplete } from './parser.js'
@@ -7,12 +6,7 @@ import { getAllProviders, safeDiscoverSessions } from './providers/index.js'
 import { claude, getClaudeConfigDirs, getDesktopSessionsDirs } from './providers/claude.js'
 import { stat } from 'node:fs/promises'
 import { aggregateProjectsIntoDays, buildPeriodDataFromDays } from './day-aggregator.js'
-import {
-  durableProjectDisplayName,
-  hasDurableProjectFilter,
-  reconcileDurableProjectDay,
-  sliceDayToProvider,
-} from './durable-project-reconciliation.js'
+import { hasDurableProjectFilter, reconcileDurableProjectDay, sliceDayToProvider } from './durable-project-reconciliation.js'
 import { aggregateModelEfficiency } from './model-efficiency.js'
 import { enrichModelsWithObservedPerformance } from './model-performance.js'
 import { aggregateModels } from './models-report.js'
@@ -26,6 +20,7 @@ import { buildGranularHistory } from './granular-history.js'
 import { isSnapshotReadMode, withReadFreshness } from './read-lifecycle.js'
 import { hydrateCopilotDailyCache } from './copilot-chat-journal-hydration.js'
 import { buildUsageBreakdowns } from './usage-breakdowns.js'
+import { friendlyProject, populateProjectRollups } from './project-report.js'
 import { buildMobileFoundationPayload } from './sharing/mobile-foundation.js'
 import { readProjectRegistry } from './project-registry.js'
 import {
@@ -560,71 +555,7 @@ export async function buildMenubarPayloadForRange(periodInfo: PeriodInfo, opts: 
     dailyHistory = [...historyFromCache, ...todayFromParse]
   }
 
-  const home = homedir()
-  const friendlyFromPath = (path: string | undefined, fallback: string): string => {
-    if (!path) return fallback
-    if (path === home || path === home + '/') return 'Home'
-    return path.split('/').filter(Boolean).pop() || fallback
-  }
-  const friendlyProject = (p: ProjectSummary) => friendlyFromPath(p.projectPath || p.project, p.project)
-  const sessionDetailsOf = (p: ProjectSummary) => [...p.sessions]
-    .sort((a, b) => b.totalCostUSD - a.totalCostUSD)
-    .slice(0, 10)
-    .map(s => ({
-      cost: s.totalCostUSD,
-      savingsUSD: s.totalSavingsUSD,
-      calls: s.apiCalls,
-      inputTokens: s.totalInputTokens,
-      outputTokens: s.totalOutputTokens,
-      date: s.firstTimestamp?.split('T')[0] ?? '',
-      models: Object.entries(s.modelBreakdown)
-        .map(([name, m]) => ({ name, cost: m.costUSD, savingsUSD: m.savingsUSD }))
-        .sort((a, b) => b.cost - a.cost)
-        .slice(0, 3),
-    }))
-
-  if (cacheDaysForPeriod !== null) {
-    // Project totals come from the SAME day set as the headline, so carried
-    // days count here too. The surviving-session parse contributes only what
-    // day entries cannot: the per-session drill-down and a fresher project
-    // path. Legacy residuals without a project split are represented by the
-    // synthetic Unattributed bucket, never assigned to a real project.
-    type CachedProjectTotal = { cost: number; savingsUSD: number; sessions: number; path?: string }
-    const cachedTotals = new Map<string, CachedProjectTotal>()
-    for (const d of cacheDaysForPeriod) {
-      for (const [name, p] of Object.entries(d.projects ?? {})) {
-        const acc = cachedTotals.get(name) ?? { cost: 0, savingsUSD: 0, sessions: 0 }
-        acc.cost += p.cost
-        acc.savingsUSD += p.savingsUSD
-        acc.sessions += p.sessions
-        if (!acc.path && p.path) acc.path = p.path
-        cachedTotals.set(name, acc)
-      }
-    }
-    const liveByName = new Map(scanProjects.map(p => [p.project, p]))
-    const names = new Set([...cachedTotals.keys(), ...liveByName.keys()])
-    currentData.projects = [...names].map(name => {
-      const cached = cachedTotals.get(name)
-      const live = liveByName.get(name)
-      return {
-        name: live ? friendlyProject(live) : friendlyFromPath(cached?.path, durableProjectDisplayName(name)),
-        cost: cached?.cost ?? live!.totalCostUSD,
-        savingsUSD: cached?.savingsUSD ?? live!.totalSavingsUSD,
-        // max for the same reason as the headline: start-day bucketing vs
-        // active-day counting, both lower bounds of distinct sessions.
-        sessions: Math.max(cached?.sessions ?? 0, live?.sessions.length ?? 0),
-        ...(live ? { sessionDetails: sessionDetailsOf(live) } : {}),
-      }
-    }).sort((a, b) => b.cost - a.cost)
-  } else {
-    currentData.projects = scanProjects.map(p => ({
-      name: friendlyProject(p),
-      cost: p.totalCostUSD,
-      savingsUSD: p.totalSavingsUSD,
-      sessions: p.sessions.length,
-      sessionDetails: sessionDetailsOf(p),
-    }))
-  }
+  populateProjectRollups(currentData, scanProjects, cacheDaysForPeriod)
 
   const effMap = aggregateModelEfficiency(scanProjects)
   currentData.modelEfficiency = [...effMap.entries()].map(([name, eff]) => ({
